refact: refactor error handling

+++ src/controllers/users/deleteUser.js
@@ -2,7 +2,11 @@ const UserRepository = require('../../repository/users');
 
 module.exports = {
     deleteUser: async (id) => {
-        const message = await UserRepository.removeUser(id);
-        return message;
+        const user = await UserRepository.selectById(id);
+        if(!user) {
+            throw new Error('USER_DOES_NOT_EXIST');
+        };
+        const result = await UserRepository.removeUser(id);
+        return result;
     },
 };
